feat: move files into server subdirectory
