+++ server/src/model.js
@@ -1,0 +1,14 @@
+const { Pool } = require('pg');
+
+const { PG_URI } = process.env;
+
+const pool = new Pool({
+  connectionString: PG_URI,
+});
+
+module.exports = {
+  query: (text, params, callback) => {
+    console.log('executed query', text);
+    return pool.query(text, params, callback);
+  },
+};
